+++ bld/wv/linux/c/lnxinit.c
@@ -146,7 +146,7 @@ bool TBreak()
     return( ret );
 }
 
-long _fork( char *cmd, unsigned len )
+long _fork( char *cmd, size_t len )
 {
     char    buff[256];
     const char    *argv[4];
