+++ bld/wv/rdos/c/rdosinit.c
@@ -118,7 +118,7 @@ void RestoreHandlers()
 {
 }
 
-long _fork( char *cmd, unsigned len )
+long _fork( char *cmd, size_t len )
 {
     cmd = cmd; len = len;
     return( 0 );
